Rename and restructure files and folders.
